+++ src/Plugin/Field/FieldFormatter/WidgetFormatter.php
@@ -2,12 +2,10 @@
 
 namespace Drupal\opening_hours\Plugin\Field\FieldFormatter;
 
-use Drupal\Core\Config\ImmutableConfig;
 use Drupal\Core\Field\FieldDefinitionInterface;
 use Drupal\Core\Field\FormatterBase;
 use Drupal\Core\Field\FieldItemListInterface;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\Language\LanguageManagerInterface;
 use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
 use Drupal\Core\StringTranslation\TranslationInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -26,20 +24,6 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 class WidgetFormatter extends FormatterBase implements ContainerFactoryPluginInterface {
 
   /**
-   * The opening hours configuration.
-   *
-   * @var \Drupal\Core\Config\ImmutableConfig
-   */
-  protected $openingHoursConfig;
-
-  /**
-   * The language manager.
-   *
-   * @var \Drupal\Core\Language\LanguageManagerInterface
-   */
-  protected $languageManager;
-
-  /**
    * Constructs a EntranceFeeFormatter object.
    *
    * @param string $plugin_id
@@ -56,10 +40,6 @@ class WidgetFormatter extends FormatterBase implements ContainerFactoryPluginInt
    *   The view mode.
    * @param array $third_party_settings
    *   Any third party settings.
-   * @param \Drupal\Core\Config\ImmutableConfig $opening_hours_config
-   *   The Opening hours configuration.
-   * @param \Drupal\Core\Language\LanguageManagerInterface $language_manager
-   *   The language manager to get the current language from.
    * @param \Drupal\Core\StringTranslation\TranslationInterface $translation
    *   The String translation.
    */
@@ -71,8 +51,6 @@ class WidgetFormatter extends FormatterBase implements ContainerFactoryPluginInt
     $label,
     $view_mode,
     array $third_party_settings,
-    ImmutableConfig $opening_hours_config,
-    LanguageManagerInterface $language_manager,
     TranslationInterface $translation
   ) {
     parent::__construct(
@@ -85,8 +63,6 @@ class WidgetFormatter extends FormatterBase implements ContainerFactoryPluginInt
       $third_party_settings
     );
 
-    $this->openingHoursConfig = $opening_hours_config;
-    $this->languageManager = $language_manager;
     $this->setStringTranslation($translation);
   }
 
@@ -105,14 +81,6 @@ class WidgetFormatter extends FormatterBase implements ContainerFactoryPluginInt
    * @return static
    */
   public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    /* @var $openingHoursConfig \Drupal\Core\Config\ImmutableConfig */
-    $openingHoursConfig = $container
-      ->get('config.factory')
-      ->get('opening_hours.settings');
-
-    /* @var $languageManager \Drupal\Core\Language\LanguageManagerInterface */
-    $languageManager = $container->get('language_manager');
-
     /* @var $stringTranslation \Drupal\Core\StringTranslation\TranslationInterface */
     $stringTranslation = $container->get('string_translation');
 
@@ -124,8 +92,6 @@ class WidgetFormatter extends FormatterBase implements ContainerFactoryPluginInt
       $configuration['label'],
       $configuration['view_mode'],
       $configuration['third_party_settings'],
-      $openingHoursConfig,
-      $languageManager,
       $stringTranslation
     );
   }
@@ -138,23 +104,12 @@ class WidgetFormatter extends FormatterBase implements ContainerFactoryPluginInt
 
     foreach ($items as $delta => $item) {
       $element[$delta] = [
-        '#theme' => 'opening_hours_widget',
-        '#type' => $this->getSetting('widget_type'),
+        '#type' => 'opening_hours_widget',
+        '#widget_type' => $this->getSetting('widget_type'),
         '#service_id' => $item->service,
         '#channel_id' => $item->channel,
-        '#language' => $this->languageManager->getCurrentLanguage()->getId(),
       ];
     }
-
-    // Attach widget + endpoint configuration.
-    $element['#attached']['library'][] = 'opening_hours/widget';
-    $element['#attached']['drupalSettings']['openingHours']['endpoint'] = $this
-      ->openingHoursConfig
-      ->get('endpoint');
-    $element['#attached']['drupalSettings']['openingHours']['language'] = $this
-      ->languageManager
-      ->getCurrentLanguage()
-      ->getId();
 
     return $element;
   }