+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All Notable changes to `drupal/opening-hours` module.
 ### Added
 
 * DMOH-26: Added support to set widget tag data-date as "today" or "tomorrow".
+* DMOH-27: Added custom render element (opening_hours_widget) to limit the code
+  required to render the widget tag.
 
 ## [8.x-1.0-aplha2]